bug 1462749: move analyze model fetches view
