+++ webapp-django/crashstats/manage/admin_urls.py
@@ -4,11 +4,13 @@ from . import admin
 
 
 urlpatterns = [
+    url('^analyze-model-fetches/$',
+        admin.analyze_model_fetches,
+        name='analyze_model_fetches'),
     url('^crash-me-now/$',
         admin.crash_me_now,
         name='crash_me_now'),
     url('^sitestatus/$',
         admin.site_status,
         name='site_status'),
-
 ]
