--- webapp-django/crashstats/manage/jinja2/manage/analyze-model-fetches.html
@@ -1,64 +1,0 @@
-{% extends "manage/base.html" %}
-
-{% block page_title %}{{ super() }} - Analyze Model Fetches{% endblock %}
-
-{% block admin_title %}Analyze Model Fetches{% endblock %}
-
-{% block site_css %}
-{{ super() }}
-{% stylesheet 'tablesorter' %}
-{% stylesheet 'analyze-model-fetches' %}
-{% endblock %}
-
-{% block site_js %}
-{{ super() }}
-{% javascript 'tablesorter' %}
-{% javascript 'analyze-model-fetches' %}
-{% endblock %}
-
-{% block mainbody %}
-<div class="panel">
-    <div class="body">
-        <table class="data-table vertical analysis tablesorter">
-            <thead>
-                <tr>
-                    <th rowspan="2">API Class</th>
-                    <th colspan="4" class="{sorter: false}"># Uses</th>
-                    <th colspan="3" class="{sorter: false}">Times (sec)</th>
-                </tr>
-                <tr class="sort-keys">
-                    <th>Hits</th>
-                    <th>Hits%</th>
-                    <th>Misses</th>
-                    <th>Both</th>
-                    <th>Hits</th>
-                    <th>Misses</th>
-                    <th>Both</th>
-                </tr>
-            </thead>
-            <tbody>
-                {% for item, info in records %}
-                <tr class="{{ loop.cycle('odd', 'even') }}">
-                    <td>{{ truncatechars(item, 150) }}</td>
-                    <td>{{ info['uses']['hits'] }}</td>
-                    <td>{{ info['uses']['hits_percentage'] }}</td>
-                    <td>{{ info['uses']['misses'] }}</td>
-                    <td>{{ info['uses']['both'] }}</td>
-                    <td>{{ info['times']['hits'] | msec2sec }}</td>
-                    <td>{{ info['times']['misses'] | msec2sec }}</td>
-                    <td>{{ info['times']['both'] | msec2sec }}</td>
-                </tr>
-                {% endfor %}
-            </tbody>
-        </table>
-
-
-        <p><b>How This Works</b></p>
-        <p>
-            Every time our Django views need data from the implementation classes, a
-            bean counter is incremented on it being used, how long it took
-            and whether or not it was able to benefit from the cache.
-        </p>
-    </div>
-</div>
-{% endblock %}
