+++ webapp-django/crashstats/manage/jinja2/manage/home.html
@@ -24,14 +24,6 @@
                 Super Search Fields
             </a></li>
         </ul>
-
-        <h3>Tools</h3>
-        <ul>
-            <li><a href="{{ url('manage:analyze_model_fetches') }}">
-                <span class="fa fa-dashboard fa-5x"></span>
-                Analyze Model Fetches
-            </a></li>
-        </ul>
     </div>
 </div>
