+++ webapp-django/crashstats/manage/templates/admin/socorro_admin_index.html
@@ -22,6 +22,13 @@
             </caption>
             <tr>
                 <th scope="row">
+                    <a href="{% url 'siteadmin:analyze_model_fetches' %}">Analyze model fetches</a>
+                </th>
+                <td></td>
+                <td></td>
+            </tr>
+            <tr>
+                <th scope="row">
                     <a href="{% url 'siteadmin:crash_me_now' %}">Crash me now</a>
                 </th>
                 <td></td>
