+++ webapp-django/crashstats/manage/urls.py
@@ -8,9 +8,6 @@ urlpatterns = [
     url('^$',
         views.home,
         name='home'),
-    url('^analyze-model-fetches/$',
-        views.analyze_model_fetches,
-        name='analyze_model_fetches'),
     url('^graphics-devices/$',
         views.graphics_devices,
         name='graphics_devices'),
