+++ webapp-django/crashstats/manage/views.py
@@ -1,9 +1,7 @@
-import hashlib
 from builtins import str
 
 from django import http
 from django.contrib import messages
-from django.core.cache import cache
 from django.shortcuts import redirect, render
 
 from crashstats.crashstats.models import GraphicsDevices
@@ -19,40 +17,6 @@ from . import utils
 def home(request, default_context=None):
     context = default_context or {}
     return render(request, 'manage/home.html', context)
-
-
-@superuser_required
-def analyze_model_fetches(request):
-    context = {}
-    all_ = cache.get('all_classes') or []
-    records = []
-    for item in all_:
-        itemkey = hashlib.md5(item.encode('utf-8')).hexdigest()
-
-        data = {}
-        data['times'] = {}
-        data['times']['hits'] = cache.get('times_HIT_%s' % itemkey, 0)
-        data['times']['misses'] = cache.get('times_MISS_%s' % itemkey, 0)
-        data['times']['both'] = (
-            data['times']['hits'] + data['times']['misses']
-        )
-        data['uses'] = {}
-        data['uses']['hits'] = cache.get('uses_HIT_%s' % itemkey, 0)
-        data['uses']['misses'] = cache.get('uses_MISS_%s' % itemkey, 0)
-        data['uses']['both'] = (
-            data['uses']['hits'] + data['uses']['misses']
-        )
-        data['uses']['hits_percentage'] = (
-            data['uses']['both'] and
-            round(
-                100.0 * data['uses']['hits'] / data['uses']['both'],
-                1
-            ) or
-            'n/a'
-        )
-        records.append((item, data))
-    context['records'] = records
-    return render(request, 'manage/analyze-model-fetches.html', context)
 
 
 @superuser_required
